requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
#opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
tensorboard>=2.4.1
pandas>=1.1.4
flask>=2.01
Flask-SQLAlchemy>=2.5.1
SQLAlchemy>=1.4.36
seaborn>=0.11.2
opencv-python>=4.6.0.66
scikit-learn>=1.1.1

#pip install opencv-python
#pip install torch
#pip install seaborn
#pip install scikit-learn
